+++ webapp/src/main/java/ar/edu/itba/paw/webapp/validators/SpeciesIdMatchValidator.java
@@ -20,7 +20,7 @@ public class SpeciesIdMatchValidator implements ConstraintValidator<SpeciesIdMat
     public boolean isValid(Object value,
                            ConstraintValidatorContext context) {
 
-        Integer val = (Integer) value;
+        Long val = (Long) value;
 
         if(val == null)
             return false;
